♻️ [Refactoring]: #433 [FE] FileTree の FileNodeItem を独立ファイルへ抽出
FileTree/index.tsx 内に定義していた再帰ノード FileNodeItem を
FileNodeItem/index.tsx へ切り出し、FileTree からは import で参照する。
挙動・DOM 出力・公開 API・呼び出し元（AppSidebar）は不変の純内部リファクタ。

Refs #433
